gsl_api/requirements.txt
Django>=2.0
django-rest-framework
gunicorn
